Move calculateFormQuarterFromDate to its own file
